refactoring: rename test folder.
