+++ integration-test/src/tests/mod.rs
@@ -1,0 +1,5 @@
+mod atomic_swap_ckb_to_udt;
+mod atomic_swap_rpc;
+pub(crate) mod atomic_swap_udt_to_udt;
+mod rpc;
+pub(crate) mod small_blank_check;
