+++ integration-test/src/tests/small_blank_check.rs
