--- integration-test/src/tests_otx_sighash_lock/mod.rs
@@ -1,4 +1,0 @@
-pub(crate) mod atomic_swap;
-mod atomic_swap_rpc;
-mod rpc;
-pub(crate) mod small_blank_check;
